+++ WebApplication_Assignment SkillsLab2023/DataTransferObjects/EnrolmentDTO.cs
@@ -8,8 +8,8 @@ namespace WebApplication_Assignment_SkillsLab2023.DataTransferObjects
 {
     public class EnrolmentDTO
     {
-        public int TrainingId {get;set;}
-        public int UserId {get;set; }
+        public byte TrainingId {get;set;}
+        public byte UserId {get;set; }
         public List<HttpPostedFileBase> Files { get;set;}
     }
 }
