+++ WebApplication_Assignment SkillsLab2023/Models/Others/CustomerModel.cs
@@ -7,7 +7,7 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
 {
     public class CustomerModel
     {
-        public int _customerId { get; set; }
+        public byte _customerId { get; set; }
         public string _customerFirstName { get; set; }
         public string _customerLastName { get; set;}
         public string _customerEmail { get; set;}
