+++ WebApplication_Assignment SkillsLab2023/Models/TrainingRegistrationModel.cs
@@ -9,9 +9,9 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
     [Table("TrainingRegistration")]
     public class TrainingRegistrationModel
     {
-        public int RegistrationId { get; set; }
-        public int TrainingID { get; set;}
-        public int UserId { get; set; }
+        public byte RegistrationId { get; set; }
+        public byte TrainingID { get; set;}
+        public byte UserId { get; set; }
         public string RegistrationStatus { get; set; }
         public bool ManagerApproval { get; set; }
