+++ WebApplication_Assignment SkillsLab2023/Models/UserModel.cs
@@ -8,11 +8,11 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
     public class UserModel
     {
         [Key]
-        public int UserId { get; set; }
+        public byte UserId { get; set; }
         public string NIC { get; set; }
         public string UserFirstName { get; set; }
         public string UserLastName { get; set; }
-        public int DepartmentId {  get; set; }
+        public byte DepartmentId {  get; set; }
         public string MobileNum {  get; set; }
         public string Role { get; set; }
     }
