+++ WebApplication_Assignment SkillsLab2023/Models/UserPrerequisiteModel.cs
@@ -9,9 +9,9 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
     [Table("UserPrerequisite")]
     public class UserPrerequisiteModel
     {
-        public int PrerequisiteId { get; set; }
-        public int TrainingId { get; set; }
-        public int UserId{ get; set; }
+        public byte PrerequisiteId { get; set; }
+        public byte TrainingId { get; set; }
+        public byte UserId { get; set; }
         public string PrerequisiteDescription { get; set; }
         public string FilePath { get; set; }
     }
