+++ WebApplication_Assignment SkillsLab2023/Models/UserRolesModel.cs
@@ -7,9 +7,9 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
 {
     public class UserRolesModel
     {
-        public int UserRoleId { get; set; }
-        public int UserId { get; set; }
-        public int RoleId { get; set; }
+        public byte UserRoleId { get; set; }
+        public byte UserId { get; set; }
+        public byte RoleId { get; set; }
         public string RoleName {  get; set; }
     }
 }
